requirements.txt
antlr4-python3-runtime==4.9.3
anyio==4.3.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
Brlapi==0.8.3
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
evdev==1.4.0
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.14.1
fastjsonschema==2.19.1
filelock==3.13.3
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.3.1
future==0.18.2
gdown==5.1.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
hydra-core==1.3.2
idna==3.3
imagecodecs==2024.1.1
imageio==2.34.1
imglyb==2.1.0
importlib-metadata==4.6.4
input-remapper==2.0.0
ipykernel==6.29.4
ipython==8.24.0
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
jgo==1.0.5
Jinja2==3.1.3
JPype1==1.5.0
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
kaggle==1.6.8
keyring==23.5.0
kiwisolver==1.4.5
labeling==0.1.13
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.2.1
notebook_shim==0.2.4
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
opencv-python==4.9.0.80
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pillow==10.3.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.8.2
pydbus==0.6.0
Pygments==2.17.2
PyGObject==3.42.1
pyimagej==1.4.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.2
referencing==0.35.0
reportlab==3.6.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
s-tui==1.1.3
scikit-image==0.23.2
scipy==1.12.0
screen-resolution-extra==0.0.0
scyjava==1.9.1
SecretStorage==3.3.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
systemd-python==234
terminado==0.18.1
text-unidecode==1.3
texttable==1.6.4
tifffile==2024.2.12
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
triton==2.2.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
urwid==2.1.2
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
xarray==2024.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
